requirements.txt
asgiref==3.8.1
Django==5.2
django-formtools==2.5.1
django-otp==1.6.0
django-phonenumber-field==8.1.0
django-two-factor-auth==1.17.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
PyJWT==2.9.0
pypng==0.20220715.0
qrcode==7.4.2
sqlparse==0.5.3
typing_extensions==4.13.2
